Add new office decor assets and placement notes
new office decor. I was not sure what was possible. SO I gave some options. I am also adding some new suggestions on where to put things and the "feel". The QR code leads to my linked in. I plan to make one for this portfolio later. On the frames reuse them more than once along with the CTA button.